requirements.txt
streamlit==1.28.0
pandas==2.1.0
psycopg2-binary==2.9.9
Faker==19.13.0
fpdf2==2.7.9
graphviz==0.20.1
watchdog